requirements.txt
Flask
tensorflow
Pillow
numpy